simplication du code pour récupérer les images sentinel en précisant les bandes et filtre sur lesnuages en direct, sans filtre sur les ombres

+++ src/utils.py
@@ -3,6 +3,7 @@ from pathlib import Path
 import s3fs
 import PIL
 from tqdm import tqdm
+from shapely.geometry import box
 
 from astrovision.data.satellite_image import (
     SatelliteImage,
@@ -27,6 +28,7 @@ def upload_satelliteImages(
     dim,
     n_bands,
     num_poly,
+    aoi,
     check_nbands12=False,
 ):
     """
@@ -52,10 +54,11 @@ def upload_satelliteImages(
         images_paths[i] = lpath + "/" + images_paths[i]
 
     print("Lecture des images")
-    list_satellite_images = [
-        SatelliteImage.from_raster(filename, n_bands=n_bands)
-        for filename in tqdm(images_paths)
-    ]
+    list_satellite_images = []
+    for filename in tqdm(images_paths):
+        image_sat = SatelliteImage.from_raster(filename, n_bands=n_bands)
+        if image_sat.array.shape[1] >= dim and image_sat.array.shape[2] >= dim:
+            list_satellite_images.append(image_sat)
 
     print(f"Découpage des images en taille {dim}")
     splitted_list_images = [
@@ -66,26 +69,30 @@ def upload_satelliteImages(
     for i in tqdm(range(len(splitted_list_images))):
         image = splitted_list_images[i]
         bb = image.bounds
-        filename = str(int(bb[0])) + "_" + str(int(bb[1])) + "_" + str(num_poly) + "_" + str(i)
+        left, bottom, right, top = bb
+        bbox = box(left, bottom, right, top)
 
-        lpath_image = lpath + "/" + filename + ".tif"
+        if aoi.contains(bbox):
+            filename = str(int(bb[0])) + "_" + str(int(bb[1])) + "_" + str(num_poly) + "_" + str(i)
 
-        image.to_raster(lpath_image)
+            lpath_image = lpath + "/" + filename + ".tif"
 
-        if check_nbands12:
-            try:
-                image = SatelliteImage.from_raster(
-                    file_path=lpath_image,
-                    n_bands=12,
-                )
+            image.to_raster(lpath_image)
+
+            if check_nbands12:
+                try:
+                    image = SatelliteImage.from_raster(
+                        file_path=lpath_image,
+                        n_bands=12,
+                    )
+                    exportToMinio(lpath_image, rpath)
+                    os.remove(lpath_image)
+
+                except PIL.UnidentifiedImageError:
+                    print("L'image ne possède pas assez de bandes")
+            else:
                 exportToMinio(lpath_image, rpath)
                 os.remove(lpath_image)
-
-            except PIL.UnidentifiedImageError:
-                print("L'image ne possède pas assez de bandes")
-        else:
-            exportToMinio(lpath_image, rpath)
-            os.remove(lpath_image)
 
 
 def exportToMinio(lpath, rpath):
